Create (almost empty) graph submodule
This is going to be filled with core graph-related trait hierarchy.
