+++ gryf/src/core.rs
@@ -8,6 +8,7 @@ mod base;
 mod connect;
 mod create;
 mod edges;
+mod graph;
 mod neighbors;
 mod properties;
 mod vertices;
@@ -17,6 +18,7 @@ pub use base::*;
 pub use connect::*;
 pub use create::*;
 pub use edges::*;
+pub use graph::*;
 pub use neighbors::*;
 pub use properties::*;
 pub use vertices::*;
